# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Hoteles",
        "description": "API para gestionar hoteles, habitaciones, huéspedes y servicios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/service": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener lista de servicios",
                "description": "Devuelve una lista de servicios con la posibilidad de filtrar por nombre y descripción, y paginar los resultados.",
                "operationId": "a1d1d1d1937d6ed895e050186e3c1750",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre del servicio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filtrar por descripción del servicio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de servicios obtenida"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "post": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Crear un nuevo servicio",
                "description": "Crea un nuevo huésped con los datos introducidos",
                "operationId": "e2348999f2c883adc269bd7810760f1c",
                "requestBody": {
                    "description": "Datos necesarios para crear un servicio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio creado"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/service/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener un servicio",
                "operationId": "c9da8db4522a54707000d628a12b467d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio encontrado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Eliminar un servicio",
                "description": "Elimina un servicio por su ID.",
                "operationId": "37d170d40a128108cade24e799dcb661",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio eliminado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                }
            }
        },
        "/api/services/{id}": {
            "put": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Actualizar un servicio",
                "description": "Actualiza un servicio con los datos introducidos",
                "operationId": "14e625e6fca9956d1143de1f41a8a228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos que se pueden actualizar del servicio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servicio actualizado"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            }
        }
    },
...
}